feat: Reorganized AlertBanner
